Consolidate testing code into sub-module
