+++ core/src/commands.rs
@@ -13,6 +13,8 @@ use self::refresh::RefrechActionError;
 
 mod publish_datapoint;
 mod refresh;
+#[cfg(test)]
+mod test_utils;
 
 pub enum PoolCommand {
     Bootstrap,
