+++ core/src/commands/publish_datapoint.rs
@@ -134,172 +134,22 @@ fn compute_new_datapoint(datapoint: i64, old_datapoint: i64) -> i64 {
 
 #[cfg(test)]
 mod tests {
-    use std::convert::TryFrom;
     use std::convert::TryInto;
 
+    use super::*;
+    use crate::commands::test_utils::{
+        find_input_boxes, make_datapoint_box, make_pool_box, make_wallet_unspent_box,
+        OracleBoxMock, PoolBoxMock, WalletDataMock,
+    };
+    use crate::contracts::refresh::RefreshContract;
     use ergo_lib::chain::ergo_state_context::ErgoStateContext;
-    use ergo_lib::chain::transaction::unsigned::UnsignedTransaction;
-    use ergo_lib::chain::transaction::TxId;
-    use ergo_lib::chain::transaction::TxIoVec;
     use ergo_lib::ergotree_interpreter::sigma_protocol::private_input::DlogProverInput;
     use ergo_lib::ergotree_ir::chain::address::AddressEncoder;
     use ergo_lib::ergotree_ir::chain::ergo_box::box_value::BoxValue;
-    use ergo_lib::ergotree_ir::chain::ergo_box::ErgoBox;
-    use ergo_lib::ergotree_ir::chain::ergo_box::NonMandatoryRegisterId;
-    use ergo_lib::ergotree_ir::chain::ergo_box::NonMandatoryRegisters;
-    use ergo_lib::ergotree_ir::chain::token::Token;
-    use ergo_lib::ergotree_ir::chain::token::TokenId;
-    use ergo_lib::ergotree_ir::ergo_tree::ErgoTree;
-    use ergo_lib::ergotree_ir::mir::constant::Constant;
-    use ergo_lib::ergotree_ir::mir::expr::Expr;
-    use ergo_lib::ergotree_ir::sigma_protocol::dlog_group::EcPoint;
-    use ergo_lib::ergotree_ir::sigma_protocol::sigma_boolean::ProveDlog;
+    use ergo_lib::ergotree_ir::chain::token::{Token, TokenId};
     use ergo_lib::wallet::signing::TransactionContext;
     use ergo_lib::wallet::Wallet;
-    use ergo_node_interface::node_interface::NodeError;
     use sigma_test_util::force_any_val;
-
-    use crate::box_kind::OracleBoxWrapper;
-    use crate::box_kind::PoolBoxWrapper;
-    use crate::contracts::oracle::OracleContract;
-    use crate::contracts::pool::PoolContract;
-    use crate::contracts::refresh::RefreshContract;
-    use crate::oracle_state::StageError;
-
-    use super::*;
-
-    #[derive(Clone)]
-    struct PoolBoxMock {
-        pool_box: PoolBoxWrapper,
-    }
-
-    impl PoolBoxSource for PoolBoxMock {
-        fn get_pool_box(&self) -> std::result::Result<PoolBoxWrapper, StageError> {
-            Ok(self.pool_box.clone())
-        }
-    }
-
-    #[derive(Clone)]
-    struct OracleBoxMock {
-        oracle_box: OracleBoxWrapper,
-    }
-
-    impl LocalDatapointBoxSource for OracleBoxMock {
-        fn get_local_oracle_datapoint_box(
-            &self,
-        ) -> std::result::Result<OracleBoxWrapper, StageError> {
-            Ok(self.oracle_box.clone())
-        }
-    }
-
-    #[derive(Clone)]
-    struct WalletDataMock {
-        unspent_boxes: Vec<ErgoBox>,
-    }
-
-    impl WalletDataSource for WalletDataMock {
-        fn get_unspent_wallet_boxes(&self) -> Result<Vec<ErgoBox>, NodeError> {
-            Ok(self.unspent_boxes.clone())
-        }
-    }
-
-    fn make_pool_box(
-        datapoint: i64,
-        epoch_counter: i32,
-        pool_nft_token_id: TokenId,
-        value: BoxValue,
-        creation_height: u32,
-    ) -> PoolBoxWrapper {
-        let tokens = [Token::from((
-            pool_nft_token_id.clone(),
-            1u64.try_into().unwrap(),
-        ))]
-        .into();
-        ErgoBox::new(
-            value,
-            PoolContract::new().ergo_tree(),
-            Some(tokens),
-            NonMandatoryRegisters::new(
-                vec![
-                    (NonMandatoryRegisterId::R4, Constant::from(datapoint)),
-                    (NonMandatoryRegisterId::R5, Constant::from(epoch_counter)),
-                ]
-                .into_iter()
-                .collect(),
-            )
-            .unwrap(),
-            creation_height,
-            force_any_val::<TxId>(),
-            0,
-        )
-        .unwrap()
-        .try_into()
-        .unwrap()
-    }
-
-    fn make_datapoint_box(
-        pub_key: EcPoint,
-        datapoint: i64,
-        epoch_counter: i32,
-        oracle_token_id: TokenId,
-        reward_token: Token,
-        value: BoxValue,
-        creation_height: u32,
-    ) -> ErgoBox {
-        let tokens = vec![
-            Token::from((oracle_token_id.clone(), 1u64.try_into().unwrap())),
-            reward_token,
-        ]
-        .try_into()
-        .unwrap();
-        ErgoBox::new(
-            value,
-            OracleContract::new().ergo_tree(),
-            Some(tokens),
-            NonMandatoryRegisters::new(
-                vec![
-                    (NonMandatoryRegisterId::R4, Constant::from(pub_key)),
-                    (NonMandatoryRegisterId::R5, Constant::from(epoch_counter)),
-                    (NonMandatoryRegisterId::R6, Constant::from(datapoint)),
-                ]
-                .into_iter()
-                .collect(),
-            )
-            .unwrap(),
-            creation_height,
-            force_any_val::<TxId>(),
-            0,
-        )
-        .unwrap()
-    }
-
-    fn make_wallet_unspent_box(pub_key: ProveDlog, value: BoxValue) -> ErgoBox {
-        let c: Constant = pub_key.into();
-        let expr: Expr = c.into();
-        ErgoBox::new(
-            value,
-            ErgoTree::try_from(expr).unwrap(),
-            None,
-            NonMandatoryRegisters::empty(),
-            1,
-            force_any_val::<TxId>(),
-            0,
-        )
-        .unwrap()
-    }
-
-    fn find_input_boxes(
-        tx: UnsignedTransaction,
-        available_boxes: Vec<ErgoBox>,
-    ) -> TxIoVec<ErgoBox> {
-        tx.inputs.mapped(|i| {
-            available_boxes
-                .clone()
-                .into_iter()
-                .find(|b| b.box_id() == i.box_id)
-                .unwrap()
-        })
-    }
 
     #[test]
     fn test_publish_datapoint() {
